Implement char manipulating functions & fix `size_t`
`size_t` was previously implemented as `unsigned int` now it's changed to `long unsigned int`

+++ ft_bzero.c
@@ -6,14 +6,14 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/15 20:47:53 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/15 20:49:16 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 14:23:48 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, long unsigned int n)
 {
-	char			*ptr;
-	unsigned int	i;
+	char				*ptr;
+	long unsigned int	i;
 
 	i = 0;
 	ptr = s;
