+++ ft_memset.c
@@ -6,14 +6,14 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/15 20:43:16 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/15 20:46:38 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 14:23:12 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, long unsigned int n)
 {
-	char			*ptr;
-	unsigned int	i;
+	char				*ptr;
+	long unsigned int	i;
 
 	i = 0;
 	ptr = s;
